Misc: fix pipeline issues

+++ src/server/game/Spells/SpellInfo.cpp
@@ -430,8 +430,8 @@ SpellEffectInfo::SpellEffectInfo(SpellInfo const* spellInfo, SpellEffectEntry co
     MiscValueB = _effect.EffectMiscValue[1];
     Mechanic = Mechanics(_effect.EffectMechanic);
     PositionFacing = _effect.EffectPosFacing;
-    TargetA = SpellImplicitTargetInfo(_effect.ImplicitTarget[0] == 0 ? TARGET_UNIT_CASTER : _effect.ImplicitTarget[0]);
-    TargetB = SpellImplicitTargetInfo(_effect.ImplicitTarget[1] == 0 ? TARGET_UNIT_CASTER : _effect.ImplicitTarget[1]);
+    TargetA = SpellImplicitTargetInfo(_effect.ImplicitTarget[0] == 0 ? uint32(TARGET_UNIT_CASTER) : uint32(_effect.ImplicitTarget[0]));
+    TargetB = SpellImplicitTargetInfo(_effect.ImplicitTarget[1] == 0 ? uint32(TARGET_UNIT_CASTER) : uint32(_effect.ImplicitTarget[1]));
     TargetARadiusEntry = sSpellRadiusStore.LookupEntry(_effect.EffectRadiusIndex[0]);
     TargetBRadiusEntry = sSpellRadiusStore.LookupEntry(_effect.EffectRadiusIndex[1]);
     ChainTargets = _effect.EffectChainTargets;
